pyproject.toml
[tool.poetry]
name = "photoshare-project"
version = "0.1.0"
description = ""
authors = ["Mykyta-Harashchenko <garasenkonikita538@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.10"
fastapi = "^0.111.1"
faker = "^26.0.0"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
sqlalchemy = "^2.0.31"
uvicorn = "^0.30.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic-settings = "^2.4.0"
redis = "^5.0.8"
fastapi-mail = "^1.4.1"
bcrypt = "3.1.7"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.41.0"
sphinx = "^8.0.2"
trelent = "^1.0"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
pytest-cov = "^5.0.0"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.9.0"
qrcode = "^7.4.2"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"





